# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: osu! Mappool Analyzer API
  version: "1.0.0"
  description: >
    Exposes the Tournament aggregate (configuration), the Beatmap aggregate
    (import), and the Analysis Engine's output (Analyses and Reports).
    Analyses and Reports are computed on demand from the current Tournament
    state — they are derived data, never stored independently of the
    Analyses they were built from (docs/04-architecture-principles.md
    Principle 5; docs/13-report-specification.md).
  license:
    name: UNLICENSED
    url: https://github.com/Ascaveth/osu-mappool-analyzer
servers:
  - url: https://api.mappool-analyzer.example/v1
    description: Production (placeholder — no deployment exists yet)
...
paths:
  /tournaments:
    get:
      operationId: listTournaments
      tags: [Tournaments]
      summary: List tournaments
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: q
          in: query
          description: Case-insensitive substring match against tournament name.
          schema: { type: string }
        - name: sort
          in: query
          description: Sort key. Prefix with `-` for descending.
          schema: { type: string, enum: [name, -name], default: name }
      responses:
        "200":
          description: Paginated list of tournaments (summary form — no stages/categories/slots).
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/TournamentSummary" }
                  pagination: { $ref: "#/components/schemas/CursorPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
    post:
      operationId: createTournament
      tags: [Tournaments]
      summary: Create a tournament from a configuration
      description: >
        Creates the Tournament aggregate (Tournament → Stage → Category →
        Slot) in one transactionally-consistent write, per
        docs/06-domain-model.md's aggregate boundary. Slots are generated
        from each category's `slotCount`, unfilled (`beatmap_id` null).
        Matches the configuration shape in
        docs/07-tournament-configuration.md.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TournamentConfiguration" }
            examples:
              qualifiers_only:
                summary: A single-stage tournament with five mod categories
                value:
                  name: "Example Open"
                  edition: "2026"
                  stages:
                    - name: "Qualifiers"
                      order: 1
                      categories:
                        - { name: "NM", order: 1, slotCount: 5 }
                        - { name: "HD", order: 2, slotCount: 2 }
                        - { name: "HR", order: 3, slotCount: 2 }
                        - { name: "DT", order: 4, slotCount: 2 }
                        - { name: "FM", order: 5, slotCount: 2 }
      responses:
        "201":
          description: Tournament created.
          headers:
            Location: { schema: { type: string, format: uri } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tournament" }
              examples:
                created:
                  value:
                    id: "8f14e1f0-9b1a-4e6a-9c0a-1f2e3d4c5b6a"
                    name: "Example Open"
                    edition: "2026"
                    stages:
                      - id: "1b2c3d4e-5f60-4718-9a0b-1c2d3e4f5061"
                        name: "Qualifiers"
                        order: 1
                        categories:
                          - id: "2c3d4e5f-6071-4829-ab1c-2d3e4f506172"
                            name: "NM"
                            order: 1
                            slots:
                              - { id: "3d4e5f60-7182-4930-bc2d-3e4f50617283", position: 1, beatmap_id: null }
        "400": { $ref: "#/components/responses/BadRequest" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /tournaments/{tournamentId}:
    parameters:
      - $ref: "#/components/parameters/TournamentId"
    get:
      operationId: getTournament
      tags: [Tournaments]
      summary: Get a tournament's full structure
      description: >
        Returns the full Tournament → Stage → Category → Slot tree.
        Slots reference Beatmaps by `beatmap_id` only, never embedded —
        Beatmap is a separate aggregate (docs/06-domain-model.md). Fetch
        `/beatmaps/{beatmapId}` for beatmap detail.
      responses:
        "200":
          description: Tournament found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tournament" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
    patch:
      operationId: updateTournament
      tags: [Tournaments]
      summary: Rename a tournament or change its edition
      description: >
        Limited to `name` and `edition`. Structural changes (stages,
        categories, slot counts) are not supported via PATCH in v1 — the
        Tournament aggregate is edited together as one unit
        (docs/06-domain-model.md), and a partial structural PATCH would
        require conflict rules (e.g. what happens to a stage's filled
        slots when its category's slotCount shrinks) that are out of
        scope until a real need for re-configuration after import
        emerges. To restructure, create a new tournament.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1 }
                edition: { type: string }
              minProperties: 1
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tournament" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
    delete:
      operationId: deleteTournament
      tags: [Tournaments]
      summary: Delete a tournament
      responses:
        "204": { description: Deleted. }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /categories/{categoryId}:
    parameters:
      - $ref: "#/components/parameters/CategoryId"
    get:
      operationId: getCategory
      tags: [Categories]
      summary: Get a category
      responses:
        "200":
          description: Category found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
components:
  parameters:
    TournamentId:
      name: tournamentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      description: Opaque pagination cursor from a previous response's `pagination.next_cursor`.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
  schemas:
...
    CursorPage:
      type: object
      required: [next_cursor, has_more]
      properties:
        next_cursor:
          type: [string, "null"]
          description: Opaque cursor for the next page, or null if there isn't one.
        has_more: { type: boolean }
    TournamentConfiguration:
      type: object
      required: [name, stages]
      properties:
        name: { type: string, minLength: 1 }
        edition: { type: string }
        stages:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/StageConfiguration" }
    StageConfiguration:
      type: object
      required: [name, order, categories]
      properties:
        name: { type: string, minLength: 1 }
        order: { type: integer, minimum: 1 }
        categories:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/CategoryConfiguration" }
        projectedStarRating:
          type: [number, "null"]
          minimum: 0
          description: >
            Optional organizer-entered target Star Rating for this stage.
            Omitted or null means unset — the Stage resource's
            `projected_star_rating` then falls back to the stage's NM1
            beatmap's star rating instead.
    CategoryConfiguration:
      type: object
      required: [name, order, slotCount]
      properties:
        name: { type: string, minLength: 1 }
        order: { type: integer, minimum: 1 }
        slotCount: { type: integer, minimum: 1 }
    TournamentSummary:
      type: object
      required: [id, name, stage_count]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        name: { type: string }
        edition: { type: string }
        stage_count: { type: integer, readOnly: true }
    Tournament:
      type: object
      required: [id, name, stages]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        name: { type: string }
        edition: { type: string }
        stages:
          type: array
          items: { $ref: "#/components/schemas/Stage" }
    Stage:
      type: object
      required: [id, name, order, categories]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        name: { type: string }
        order: { type: integer }
        categories:
          type: array
          items: { $ref: "#/components/schemas/Category" }
        projected_star_rating:
          type: [number, "null"]
          readOnly: true
          description: >
            The organizer's explicit projectedStarRating if one was set at
            creation, otherwise the stage's NM1 beatmap's star_rating if
            that slot is filled. Null if neither is available.
    Category:
      type: object
      required: [id, name, order, slots]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        name: { type: string }
        order: { type: integer }
        slots:
          type: array
          items: { $ref: "#/components/schemas/Slot" }
    Slot:
      type: object
      required: [id, position]
      description: >
        beatmap_id is the only beatmap reference — Slot never embeds a
        full Beatmap (docs/06-domain-model.md: Beatmap is a separate
        aggregate, referenced by ID only).
      properties:
        id: { type: string, format: uuid, readOnly: true }
        position: { type: integer }
        beatmap_id:
          type: [string, "null"]
          format: uuid
          description: Null when the slot is unfilled — a valid, expected pool state (docs/06-domain-model.md).
        effective_difficulty:
          type: [object, "null"]
          readOnly: true
          description: >
            The slot's beatmap AR/OD/CS/HP/BPM/length as they actually play
            under this slot's Category's fixed mod (e.g. HR scales
            CS/AR/OD/HP, DT scales BPM/length and rescales AR/OD's timing
            windows) — not the beatmap's raw .osu values. Null when the
            slot is unfilled, or its Category has no single fixed mod
            (FreeMod, Tiebreaker, or an unrecognized name).
          properties:
            ar: { type: number }
            od: { type: number }
            cs: { type: number }
            hp: { type: number }
            bpm: { type: number }
            length_seconds: { type: integer }
            star_rating:
              type: [number, "null"]
              description: >
                Real, mod-specific Star Rating fetched from the osu! API at
                import time (internal/enrich) for this slot's fixed mod —
                unlike the other fields here, not something this arithmetic
                transform can compute itself. Null when star rating
                fetching is disabled or this mod combination hasn't been
                enriched yet.
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
      examples:
        - type: "https://docs.mappool-analyzer.example/errors/validation-error"
          title: "Validation Error"
          status: 422
          detail: "stages[0].categories must contain at least 1 item."
          instance: "https://api.mappool-analyzer.example/v1/tournaments"
          errors:
            - { field: "stages[0].categories", message: "must contain at least 1 item" }
  responses:
    BadRequest:
      description: Malformed request (bad query params, invalid JSON body shape).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationError:
      description: Well-formed request that fails domain validation (e.g. a TournamentConfiguration violating docs/07's schema).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
      description: >
        Defined for deployments that need access control. v1 does not
        enforce it globally (see docs/14-api-specification.md#authentication) —
        appropriate for the project's current self-hosted, single-tenant
        usage; enabling it later is additive, not a breaking change. The
        empty alternative in the global `security` requirement below is
        what makes it optional rather than required: a deployment that
        wants to require it removes that alternative, not the scheme.
security:
  - {}
  - BearerAuth: []